evaluation/requirements.txt
# Dependências para avaliação de agentes
ragas>=0.1.7
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
jsonlines>=4.0.0
scikit-learn>=1.3.0
plotly>=5.17.0
tabulate>=0.9.0
